MVClized sensor data
